2 - tarefas/requirements.txt
Django==5.2
djangorestframework==3.15.0
psycopg2==2.9.9